+++ lib/hadar.rb
@@ -12,6 +12,7 @@ require_relative "hadar/slide"
 require_relative "hadar/theme"
 require_relative "hadar/deck"
 require_relative "hadar/renderer"
+require_relative "hadar/slide_list"
 
 module Hadar
   class Error < StandardError; end